Added some validation and fixed some small issues in the follower service

+++ follower-service/src/main/java/com/myfoodie/followerservice/service/FollowerServiceImpl.java
@@ -3,6 +3,8 @@ package com.myfoodie.followerservice.service;
 import com.myfoodie.followerservice.dto.FavoriteDishResponse;
 import com.myfoodie.followerservice.dto.FollowerRequest;
 import com.myfoodie.followerservice.dto.FollowerResponse;
+import com.myfoodie.followerservice.exception.FollowerExistsException;
+import com.myfoodie.followerservice.exception.UserNotFoundException;
 import com.myfoodie.followerservice.feign.FavoriteDishServiceClient;
 import com.myfoodie.followerservice.feign.UserServiceClient;
 import com.myfoodie.followerservice.model.Follower;
@@ -13,16 +15,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class FollowerServiceImpl implements FollowerService{
+public class FollowerServiceImpl implements FollowerService {
 
     private final FollowerRepository followerRepository;
     private final FavoriteDishServiceClient favoriteDishServiceClient;
+    private final UserServiceClient userServiceClient;
 
 
     @Override
@@ -52,18 +53,42 @@ public class FollowerServiceImpl implements FollowerService{
                 .followedDate(followerRequest.getFollowedDate())
                 .build();
 
+        if (!userServiceClient.checkIfUserExists(follower.getFollowerId())) {
+            throw new UserNotFoundException("The user with ID " + follower.getFollowerId() + " does not exists");
+        }
+
+        if (!userServiceClient.checkIfUserExists(follower.getUserId())) {
+            throw new UserNotFoundException("The user with ID " + follower.getUserId() + " does not exists");
+        }
+
+        if (followerRepository.existsByFollowerIdAndUserId(follower.getFollowerId(), follower.getUserId())) {
+            throw new FollowerExistsException("The follower with ID " + follower.getFollowerId() + " is already following the user with ID " + follower.getUserId());
+        }
+
+        if (follower.getFollowerId().equals(follower.getUserId())) {
+            throw new FollowerExistsException("The user can not follow it self");
+        }
+
         followerRepository.save(follower);
 
-        log.info("Follower id {} is saved", follower.getId());
+        log.info("Follower ID {} is now following user ID {}", follower.getFollowerId(), follower.getUserId());
     }
 
 
     @Override
     @Transactional
-    public void deleteFollowerByIds(String followedId,String userId) {
-        followerRepository.deleteByFollowerIdAndUserId(followedId, userId);
+    public void deleteFollowerByIds(String followedId, String userId) {
+        if (!userServiceClient.checkIfUserExists(followedId)) {
+            throw new UserNotFoundException("The user with ID " + followedId + " does not exists");
+        }
 
-        log.info("User id {} is no longer following user id {}", userId,followedId);
+        if (!userServiceClient.checkIfUserExists(userId) ){
+            throw new UserNotFoundException("The user with ID " + userId + " does not exists");
+        }
+
+        if (followerRepository.existsByFollowerIdAndUserId(followedId, userId)) {
+            followerRepository.deleteByFollowerIdAndUserId(followedId, userId);
+            log.info("User id {} is no longer following user id {}", userId, followedId);
+        }
     }
-
 }
